first test file all db inputs correct
the inputs for the db are correct for all inputs, including primEx
assEx and atnyAgntFirm.  I only ran it on file
US_7469381_B2_fulltext.html and  htmlonlyviewUS_5612524_A
regulatory.html which is one for regulatory information of a different
patent. I need to check it on a 5612524 full text and 7469381
regulatory info.