fixed url, no more whitespace in url
the list of publication numbers has a whitespace, so I had to strip the
white space off. I thought I was doing that earlier but I was not
changing the variable. So I created a new variable and had pubNum = new
variable.strip(). I used to just have pubNum.strip() and then called
the pubNum variable but that pubNum variable is not defined with the
strip function so it has no effect.